% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRlearner.R
\name{DRlearner}
\alias{DRlearner}
\title{DR-Learner, implemented with SuperLearner}
\usage{
DRlearner(
  X,
  A,
  Y,
  SL_OR = "SL.mean",
  SL_PS = "SL.mean",
  SL_CATE = "SL.mean",
  cvControl = list(V = 3L)
)
}
\arguments{
\item{X}{A data.frame of predictors.}

\item{A}{Numeric vector of treatment assignments (0, 1 only).}

\item{Y}{Numeric vector of outcome variable.}

\item{SL_OR}{A character vector containing the SuperLearner library for ensemble estimation of the first-stage outcome regression.}

\item{SL_PS}{A character vector containing the SuperLearner library for ensemble estimation of the first-stage propensity score.}

\item{SL_CATE}{A character vector containing the SuperLearner library for ensemble estimation of the second-stage regression of the double robust pseudo outcomes.}

\item{cvControl}{A list containing additional parameters for the SuperLearner cross-validation (see \code{help(SuperLearner)} for more).}
}
\value{

}
\description{
DR-Learner, implemented with SuperLearner
}
\examples{
library(SuperLearner)
SL.lib <- "SL.glmnet"
n = 100; p = 5
set.seed(1234)
X = data.frame(matrix(runif(n*p,-3,3), n, p))
A = rbinom(n, 1, 0.5)
Y = pmax(X[,1], 0) * A + X[,2] + pmin(X[,3], 0) + rnorm(n)

fit <- DRlearner(X=X, A=A, Y=Y, SL_OR = SL.lib, SL_PS = SL.lib, SL_CATE = SL.lib,
                 cvControl = list(V=5))
}
\references{
\insertRef{kennedy2020optimal}{SL.wrappers}
}
