% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRlearner.R
\name{predict.DRlearner}
\alias{predict.DRlearner}
\title{Predict for \code{DRlearner} objects}
\usage{
\method{predict}{DRlearner}(object, newdata)
}
\arguments{
\item{object}{A \code{Drlearner} object.}

\item{newdata}{A data.frame of new values for the predictors.}
}
\value{
A vector of predictions
}
\description{
Predict for \code{DRlearner} objects
}
\examples{
library(SuperLearner)
SL.lib <- "SL.glmnet"
n = 100; p = 5
set.seed(1234)
X = data.frame(matrix(runif(n*p,-3,3), n, p))
A = rbinom(n, 1, 0.5)
Y = pmax(X[,1], 0) * A + X[,2] + pmin(X[,3], 0) + rnorm(n)

fit <- DRlearner(X=X, A=A, Y=Y, SL_OR = SL.lib, SL_PS = SL.lib, SL_CATE = SL.lib,
                 cvControl = list(V=5))
predict(fit, X)
}
