% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxgboost.R
\name{predict.SL.cvboost}
\alias{predict.SL.cvboost}
\title{Prediction for wrapper SL.cvboost}
\usage{
\method{predict}{SL.cvboost}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{SL.cvboost object}

\item{newdata}{Dataframe to generate predictions}

\item{...}{Unused additional arguments}
}
\value{

}
\description{
Prediction for wrapper SL.cvboost
}
\examples{
#' set.seed(1)
n = 100; p = 5

x = matrix(rnorm(n*p), n, p)
y = pmax(x[,1], 0) + x[,2] + pmin(x[,3], 0) + rnorm(n)
fit = SL.cvboost(y, data.frame(x), newX = data.frame(x),
family = gaussian(), obsWeights = NULL)
est = predict(fit$fit)
}
